+++ sparv/core/config.py
@@ -4,7 +4,7 @@ import copy
 from collections import defaultdict
 from functools import reduce
 from pathlib import Path
-from typing import Any, Dict, Optional
+from typing import Any, Dict, Optional, Union
 
 import yaml
 import yaml.scanner
@@ -34,7 +34,10 @@ _config_default = {}  # Default config
 config_structure = {
     "classes": {"_source": "core", "_cfg": Config("classes", datatype=dict)},
     "custom_annotations": {"_source": "core", "_cfg": Config("custom_annotations", datatype=list)},
-    "install": {"_source": "core", "_cfg": Config("install", datatype=list)},
+    "install": {
+        "_source": "core",
+        "_cfg": Config("install", description="List of default installers to run", datatype=list)
+    },
     PARENT: {"_source": "core", "_cfg": Config(PARENT, datatype=str)},
     MAX_THREADS: {"_source": "core", "_cfg": Config(MAX_THREADS, datatype=Dict[str, int])},
     "preload": {"_source": "core", "_cfg": Config("preload", datatype=list)},
@@ -50,7 +53,7 @@ class Unset:
     pass
 
 
-def read_yaml(yaml_file):
+def read_yaml(yaml_file: Union[str, Path]) -> dict:
     """Read YAML file and handle errors."""
     try:
         with open(yaml_file, encoding="utf-8") as f:
@@ -203,6 +206,11 @@ def get_config_description(name):
     """Get description for config key."""
     cfg = _get(name, config_structure).get("_cfg")
     return cfg.description if cfg else None
+
+
+def get_config_object(name: str) -> Optional[Config]:
+    """Get original Config object for config key."""
+    return _get(name, config_structure).get("_cfg")
 
 
 def add_config_usage(config_key, annotator):